refactor(settings): peel settings.go sections into per-topic files
- move entry-catalog, root, projects, AI, notifications, appearance,
  keybindings, labs, and about sections out of settings.go
- normalize osStat/osLstat globals into a settingsCommand DI field
  (statFile) plus direct os.Stat/os.Lstat at free-function call sites
